+++ crates/forge-runner/src/lib.rs
@@ -57,7 +57,6 @@ pub struct RunnerConfig {
     pub exit_first: bool,
     pub fuzzer_runs: u32,
     pub fuzzer_seed: u64,
-    pub print_gas_usage: bool,
 }
 
 impl RunnerConfig {
@@ -68,14 +67,12 @@ impl RunnerConfig {
         exit_first: bool,
         fuzzer_runs: u32,
         fuzzer_seed: u64,
-        print_gas_usage: bool,
     ) -> Self {
         Self {
             workspace_root,
             exit_first,
             fuzzer_runs,
             fuzzer_seed,
-            print_gas_usage,
         }
     }
 }
@@ -246,7 +243,7 @@ pub async fn run_tests_from_crate(
     while let Some(task) = tasks.next().await {
         let result = task??;
 
-        print_test_result(&result, runner_config.print_gas_usage);
+        print_test_result(&result);
 
         if let TestCaseSummary::Failed { .. } = result {
             if runner_config.exit_first {