+++ crates/forge-runner/src/printing.rs
@@ -1,7 +1,7 @@
 use crate::test_case_summary::TestCaseSummary;
 use console::style;
 
-pub(crate) fn print_test_result(test_result: &TestCaseSummary, print_gas_usage: bool) {
+pub(crate) fn print_test_result(test_result: &TestCaseSummary) {
     if let TestCaseSummary::Skipped { .. } = test_result {
         return;
     }
@@ -50,10 +50,9 @@ pub(crate) fn print_test_result(test_result: &TestCaseSummary, print_gas_usage: 
     };
 
     let mut gas_usage = String::new();
-    if print_gas_usage {
-        if let Some(result) = test_result.gas_usage() {
-            gas_usage = format!(", gas: ~{result}");
-        }
+
+    if let Some(result) = test_result.gas_usage() {
+        gas_usage = format!(", gas: ~{result}");
     }
 
     println!("{result_header} {result_name}{fuzzer_report}{block_number_message}{result_message}{gas_usage}");