+++ crates/forge/test_utils/src/running_tests.rs
@@ -24,7 +24,6 @@ pub fn run_test_case(test: &TestCase) -> Vec<TestCrateSummary> {
             false,
             256,
             12345,
-            false,
         )),
         Arc::new(RunnerParams::new(
             corelib_path(),