+++ crates/forge/tests/integration/setup_fork.rs
@@ -113,7 +113,6 @@ fn fork_aliased_decorator() {
                 false,
                 256,
                 12345,
-                false,
             )),
             Arc::new(RunnerParams::new(
                 corelib_path(),