+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 #### Added
 
-- `--print-gas-usage`, short `-g` flag to print gas usage for each test
+- print gas usage for each test
 
 ### Cast